fixed buffer offset bug

+++ LPCRenderer/shaders/pcBarebonesCompressed.vert
@@ -2,7 +2,7 @@
 
 uniform vec3 cloudOrigin;
 uniform vec3 brickSize;
-layout(location = 0) in vec3 brickIndex;
+layout(location = 0) in uvec3 brickIndex;
 layout(location = 1) in uint bufferOffset;
 layout(location = 2) in uint bufferLength;
 
@@ -16,5 +16,7 @@ void main()
 {
 	vs_out.bufferOffset = bufferOffset;
 	vs_out.bufferLength = bufferLength;
+
 	gl_Position = vec4(cloudOrigin + brickIndex * brickSize, 1);	
+
 }
